test(containment): put the invariant test where PS-204 can find its module
CI caught the one finding this PR actually added: PS-204 §2 orphan-test-file.
`tests/scitex_agent_container/test__apptainer_only_containment.py` sat at the
mirror root, whose src dir holds no `_apptainer_only_containment.py` — a test
file naming a module that does not exist. Every other finding in that audit
run (§4b free-form help, §10w unmeasurable import budget, §13 top-level
`skills`, PS-226 job name) is pre-existing and present on green develop runs.

Moved to `tests/scitex_agent_container/config/test__container_engine.py`,
which mirrors `config/_container_engine.py` exactly — the module that
DECLARES the invariant. The tests belong with the declaration rather than
with any one launch path, because the claim spans several
(`_lifecycle/_runtime_select`, the three session runtimes, `cli_pkg/send_cmds`)
and a claim verified in only one of them is the shape that let the bare-host
`sac agents send` fallback survive unnoticed.

`test__validation_container_runtime.py` is folded into it and deleted rather
than left beside it: its premise (the enum {none, apptainer}) died with the
field, and one module now has one test file. Local audit after the move:
`0 unmasked error(s)`, PS-204 gone.
